requirements.txt
fastapi==0.110.0
python-dotenv==1.0.1
python-keycloak==3.9.1
requests==2.31.0
uvicorn==0.28.0
pydantic
httpx
python-jose==3.5.0